Changed shear multiplcation to addition
When an animation is complete or we only have a single keyframe of animation
the shear timeline applies the current frame shear to the bones base shear.

When a bone hasn't definied a shear value it defaults to 0, which means this
multiplcation is always 0.

Changed multiplication to addition to match the interpolation code below